+++ source/src/main/java/com/mauja/maujaadventures/jeu/Observateur.java
@@ -1,5 +1,5 @@
 package com.mauja.maujaadventures.jeu;
 
 public interface Observateur {
-    void miseAJour(int timer);
+    void miseAJour(long timer);
 }
